Revised Compression Method (#2)
* compress may return less than target bytes

Instead of throwing an error or zero-padding, "compress" now returns the
input bytes if there are less than or equal to "target" number of them.
I think this is logical since the goal of compress is to reduce the
complexity of the digest before making it human consumable. In this case
the complexity is already low enough to proceed.

* Revised compression method

Excess bytes are now distributed amongst the compressed bytes, instead
of being dumped into the final bit as they were before.

* Compression method comments

* Updated tests

Set new correct test outputs for revised compression method. Editted comments to be within line limit.